requirements.txt
streamlit>=1.28.0
chromadb>=0.4.0,<0.5.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
pypdf>=3.17.0
youtube-transcript-api>=0.6.0
playwright>=1.40.0
rank-bm25>=0.2.2
beautifulsoup4>=4.12.0
requests>=2.31.0

